AEK Template for ISO QWERTZ Keyboards
A tweaked version of the original template for use with the AEK II in the ISO German version.
It's using the U001 font as surrogate for the Univers type of the AEK. The measurements have been fine-tuned for the AEK II.
The symbols for brightness, etc. have been moved on top and optimised. Since I've improved the grouping of the media keys of my AEK II, this is also mirrored in the template.
As help for programmers the often used symbols "[]|\{}" have been added, together with the "¿".


https://www.cufonfonts.com/font/u001